refactor(components): migrate Notification to TypeScript
