+++ src/components/Notification/index.tsx
@@ -13,25 +13,11 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import { Notification } from './Notification';
 
-import Notification from '.';
+export {
+  Icon as NotificationIcon,
+  Button as NotificationButton,
+} from './components';
 
-describe('Notification', () => {
-  /**
-   * Style tests.
-   */
-  it('should render with default styles', () => {
-    const actual = create(<Notification />);
-    expect(actual).toMatchSnapshot();
-  });
-
-  /**
-   * Accessibility tests.
-   */
-  it('should meet accessibility guidelines', async () => {
-    const wrapper = renderToHtml(<Notification />);
-    const actual = await axe(wrapper);
-    expect(actual).toHaveNoViolations();
-  });
-});
+export default Notification;
